Add keyboard shortcuts
This commit adds keyboard shortcuts for detach-tabs.
The default shortcuts are:

* `Ctrl+Space`: Detach current tab
* `Ctrl+Shift+Space`: Detach current tab and all tabs to the right
* `Ctrl+Alt+L`: Detach all tabs with the same origin from this window
* `Ctrl+Shift+L`: Detach all tabs with the same origin from all windows

NOTE: The last item doesn't seem to work in Linux/GNOME; for now you'll
need to find a different shortcut.

Shortcuts can be changed by going to "Manage Extension Shortcuts"
from the `about:addons` page. The labels on that page should be
localized.

The shortcut labels are localized using new strings instead of reusing
the context menu items, because I'm not sure the menu items would
make sense as shortcut labels in languages other than English. That's
for translators to decide, I guess!